# artifacts/api-server/src/routes/chat.ts
import { Router, type IRouter } from "express";
import crypto from "crypto";
import { ilike, or } from "drizzle-orm";
import { db, articlesTable } from "@workspace/db";
import {
  SendMessageBody,
  SendMessageResponse,
  EscalateTicketBody,
  EscalateTicketResponse,
} from "@workspace/api-zod";
import { matchIntent, scrubPII } from "../lib/intent-matcher.js";
import { FALLBACK_RESPONSES } from "../lib/responses.js";
import { generateLLMResponse } from "../lib/llm.js";

const router: IRouter = Router();

router.post("/chat", async (req, res): Promise<void> => {
  const parsed = SendMessageBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { message, sessionId } = parsed.data;
  const scrubbedMessage = scrubPII(message);
  const intentResult = matchIntent(message);

  const searchTerms = message.split(/\s+/).filter(w => w.length > 3).slice(0, 3);
  let relatedArticles: any[] = [];
  if (searchTerms.length > 0) {
    try {
      const conditions = searchTerms.map(term =>
        or(
          ilike(articlesTable.title, `%${term}%`),
          ilike(articlesTable.description, `%${term}%`)
        )!
      );
      relatedArticles = await db
        .select()
        .from(articlesTable)
        .where(or(...conditions)!)
        .limit(3);
    } catch {
      relatedArticles = [];
    }
  }

  const fallback = FALLBACK_RESPONSES[intentResult.intent] || FALLBACK_RESPONSES["general"];

  let articleContext = "";
  if (relatedArticles.length > 0) {
    articleContext = relatedArticles
      .map(a => `Title: ${a.title}\nDescription: ${a.description}\nBody: ${a.body?.substring(0, 300)}`)
      .join("\n---\n");
  }

  let reply: string;
  const llmResponse = await generateLLMResponse(scrubbedMessage, intentResult.intent, articleContext);

  if (llmResponse) {
    reply = llmResponse;
  } else {
    reply = fallback.reply;
  }

  const responseData = {
    reply,
    intent: intentResult.intent,
    trustScore: Math.round(intentResult.confidence * 100) / 100,
    isSecurityAlert: intentResult.isSecurityAlert,
    isFinancialAdvice: intentResult.isFinancialAdvice,
    sources: fallback.sources,
    ciBreakdown: `Intent confidence: ${(intentResult.confidence * 100).toFixed(1)}% | Source trust: ${relatedArticles.length > 0 ? "94.0%" : "88.0%"} | LLM: ${llmResponse ? "Enhanced" : "Fallback"}`,
    lastUpdated: new Date(),
    relatedArticles: relatedArticles.map(a => ({
      id: a.id,
      title: a.title,
      description: a.description,
      body: a.body,
      category: a.category,
      sourceUrl: a.sourceUrl,
      trustScore: a.trustScore,
      lastUpdated: a.lastUpdated,
    })),
  };

  res.json(SendMessageResponse.parse(responseData));
});

router.post("/escalate", async (req, res): Promise<void> => {
  const parsed = EscalateTicketBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { sessionId, subject, chatHistory } = parsed.data;

  const sessionHash = crypto
    .createHash("sha256")
    .update(sessionId + Date.now().toString())
    .digest("hex")
    .substring(0, 16);

  const chatTranscript = chatHistory
    .map(msg => `[${msg.role}] ${msg.content}`)
    .join("\n");

  const ticketPayload = {
    subject: subject || "Emerald Support - Escalated Conversation",
    comment: `Session ID: ${sessionId}\n\n--- Chat Transcript ---\n${chatTranscript}\n\n--- End Transcript ---\n\nThis ticket was auto-generated by the Emerald support system.`,
    priority: "urgent",
    sessionHash,
  };

  console.log("=== ZENDESK TICKET PAYLOAD ===");
  console.log(JSON.stringify({
    ticket: {
      subject: ticketPayload.subject,
      priority: ticketPayload.priority,
      sessionHash: ticketPayload.sessionHash,
      messageCount: chatHistory.length,
    }
  }, null, 2));
  console.log("=== END ZENDESK PAYLOAD ===");

  res.json(EscalateTicketResponse.parse({
    success: true,
    ticketPayload,
  }));
});

export default router;