# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emerald Support Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Article {
  id: number;
  title: string;
  description: string;
  body: string;
  category: string;
  sourceUrl: string;
  trustScore: number;
  lastUpdated: Date;
}