# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emerald Support Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";
import type { SourceReference } from "./sourceReference";

export interface ChatResponse {
  reply: string;
  intent: string;
  trustScore: number;
  isSecurityAlert: boolean;
  isFinancialAdvice: boolean;
  sources: SourceReference[];
  ciBreakdown: string;
  lastUpdated: Date;
  relatedArticles?: Article[];
}