# lib/api-zod/src/generated/types/escalateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emerald Support Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface EscalateRequest {
  sessionId: string;
  subject: string;
  chatHistory: ChatMessage[];
}