# lib/api-zod/src/generated/types/sourceReference.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emerald Support Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface SourceReference {
  title: string;
  url: string;
  type: string;
}